fixe buggs

+++ backend/src/main/java/de/mameie/backend/rest/service/CustomerService.java
@@ -1,5 +1,6 @@
 package de.mameie.backend.rest.service;
 
+import de.mameie.backend.rest.model.ModelConverter;
 import de.mameie.backend.rest.model.dto.CustomerDto;
 import de.mameie.backend.rest.model.entity.CustomerEntity;
 import de.mameie.backend.rest.repository.CustomerRepository;
@@ -7,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CustomerService {
@@ -19,11 +21,27 @@ public class CustomerService {
     }
 
     public void save(CustomerEntity customerEntity) {
-        customerRepository.save(customerEntity);
+        if (!customerRepository.existsByEmail(customerEntity.getEmail())) {
+            customerRepository.save(customerEntity);
+        }
+
     }
-    public void delete(String email){
-        customerRepository.deleteByEmail(email);
+
+    public void delete(String email) {
+        if(customerRepository.existsByEmail(email)){
+            customerRepository.deleteByEmail(email);
+        }
     }
+
+    public void update(CustomerEntity customerEntity) {
+        if(customerRepository.existsBySign(customerEntity.getSign())){
+            CustomerEntity dbEntity = customerRepository.findBySign(customerEntity.getSign());
+            customerEntity.setId(dbEntity.getId());
+            customerEntity.setPassword(dbEntity.getPassword());
+            customerRepository.save(customerEntity);
+        }
+    }
+
     public CustomerEntity getCustomer(String sign) {
         return customerRepository.findBySign(sign);
     }
